requirement.txt
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-chroma
langchain-openai
pypdf
sentence-transformers
transformers>=4.51.0
torch
tqdm
python-dotenv
fastapi
uvicorn
